bugfix(#1261): show unsupported network popover
